pyproject.toml
[tool.poetry]
name = "hw27"
version = "0.1.0"
description = ""
authors = ["SavinovaIren <irka887@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
psycopg2 = "^2.9.4"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
